Add roaring bitmaps dependency

+++ src/lib.rs
@@ -20,5 +20,5 @@ pub enum Error {
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-pub type StateId = usize;
+pub type StateId = u32;
 pub const START_STATE_ID: StateId = 0;
